requirements.txt
# Runtime dependency
requests>=2.32.0

# Development dependencies (optional)
black
ruff
mypy
pytest
pytest-asyncio
pytest-homeassistant-custom-component